+++ src/xcb-pb2.c
@@ -273,7 +273,7 @@ static dstr get_indices(void) {
 	return res;
 }
 
-static int server_cron(event_loop el, long long id, void *data) {
+static int server_cron(event_loop el, unsigned long id, void *data) {
 	dlist_iter_t iter;
 	dlist_node_t node;
 	NOT_USED(el);
